+++ src/algo/simd/search_simd_sw.c
@@ -63,7 +63,7 @@
 #define _mmxxx_setzero_si _mm_setzero_si128
 #define _mmxxx_movemask_epi8 _mm_movemask_epi8
 
-#endif
+#endif /* __AVX2__ */
 
 #ifdef SEARCH_8_BIT
 
@@ -106,7 +106,7 @@ typedef int8_t intYY_t;
 #define dbg_add_matrix_data_xxx_YY_sw dbg_add_matrix_data_128_8_sw
 #define dbg_mmxxx_print_YYs dbg_mm_print_8s
 
-#endif
+#endif /* __AVX2__ */
 
 #else // search 16 bit
 
@@ -149,9 +149,9 @@ typedef int16_t intYY_t;
 #define dbg_add_matrix_data_xxx_YY_sw dbg_add_matrix_data_128_16_sw
 #define dbg_mmxxx_print_YYs dbg_mm_print_16s
 
-#endif
+#endif /* __AVX2__ */
 
-#endif // end search
+#endif /* SEARCH_8_BIT */
 
 #ifdef DBG_COLLECT_MATRIX
 static int d_idx;
@@ -316,7 +316,7 @@ void search_YY_XXX_sw( p_sYYinfo s, p_db_chunk chunk, p_minheap heap, p_db_chunk
 
     overflow.v = _mmxxx_setzero_si();
 
-    uint8_t dseq_search_window[CDEPTH * CHANNELS];
+    uint16_t dseq_search_window[CDEPTH * CHANNELS];
 
     size_t next_id = 0;
     size_t done = 0;
