+++ tests/algo/8/test_8_simd_utilities.c
@@ -79,8 +79,8 @@ START_TEST (test_sse_simple)
 
         p_s8info s = setup_simd_util_test( "AT" );
 
-        uint8_t dseq_search_window[CDEPTH_8_BIT * CHANNELS_8_BIT_SSE];
-        memset( dseq_search_window, 0, sizeof(uint8_t) * CDEPTH_8_BIT * CHANNELS_8_BIT_SSE );
+        uint16_t dseq_search_window[CDEPTH_8_BIT * CHANNELS_8_BIT_SSE];
+        memset( dseq_search_window, 0, sizeof(uint16_t) * CDEPTH_8_BIT * CHANNELS_8_BIT_SSE );
 
         sequence_t dseq = us_prepare_sequence( "AATG", 4, 0, 0 );
 
@@ -101,8 +101,8 @@ START_TEST (test_avx_simple)
 
         p_s8info s = setup_simd_util_test( "AT" );
 
-        uint8_t dseq_search_window[CDEPTH_8_BIT * CHANNELS_8_BIT_AVX];
-        memset( dseq_search_window, 0, sizeof(uint8_t) * CDEPTH_8_BIT * CHANNELS_8_BIT_AVX );
+        uint16_t dseq_search_window[CDEPTH_8_BIT * CHANNELS_8_BIT_AVX];
+        memset( dseq_search_window, 0, sizeof(uint16_t) * CDEPTH_8_BIT * CHANNELS_8_BIT_AVX );
 
         sequence_t dseq = us_prepare_sequence( "AATG", 4, 0, 0 );
